tracker/requirements.txt
opencv-python
numpy
websockets
pyserial
